+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change Log
-
+## [2.0.0] (2022-11-26)
+Added support for custom hover information.
 ## [1.1.0] (2022-11-06)
-Initial release to VSCode marketplace
+Initial release to VSCode marketplace.